requirements.txt
dash
plotly
pandas
pyyaml
interop